use properly buildEndpoint(port) everywhere
