+++ pkg/otelcgen/consts.go
@@ -2,16 +2,16 @@ package otelcgen
 
 const (
 	OtlpGrpcPort = 4317
-	OtlpHTTPPort = 4317
+	OtlpHTTPPort = 4318
 
-	JaegerGrpcPort         = 14250
-	JaegerThriftBinaryPort = 6832
-	JaegerThriftComactPort = 6831
-	JaegerThriftHTTPPort   = 14268
+	JaegerGrpcPort          = 14250
+	JaegerThriftBinaryPort  = 6832
+	JaegerThriftCompactPort = 6831
+	JaegerThriftHTTPPort    = 14268
 
 	ZipkinPort = 9411
 
 	StatsdPort = 8125
 
-	ExtensionsHealthCeckPort = 13133
+	ExtensionsHealthCheckPort = 13133
 )
