+++ pkg/otelcgen/extensions_test.go
@@ -11,6 +11,7 @@ import (
 
 func TestNewConfigWithExtensions(t *testing.T) {
 	cfg, err := NewConfig(
+		WithPodIP("test"),
 		WithExtensions(),
 	)
 	require.NoError(t, err)
