+++ pkg/otelcgen/otelcgen.go
@@ -59,9 +59,6 @@ const (
 	ZipkinProtocol Protocol = "zipkin"
 	OtlpProtocol   Protocol = "otlp"
 	StatsdProtocol Protocol = "statsd"
-
-	StatsdDefaultEndpoint = "test"
-	ZipkinDefaultEndpoint = "test"
 )
 
 var (
@@ -125,65 +122,6 @@ func (c *Config) buildEndpoint(port uint) string {
 
 func (c *Config) buildEndpointWithoutPort() string {
 	return c.podIP
-}
-
-// func
-// receivers
-func (c *Config) buildReceiverComponent(componentID component.ID) component.Config {
-	switch componentID {
-	case OtlpID:
-		return map[string]any{"protocols": map[string]any{
-			"grpc": &ServerConfig{TLSSetting: c.buildTLSSetting(), Endpoint: "test:4317"},
-			"http": &ServerConfig{TLSSetting: c.buildTLSSetting(), Endpoint: "test:4318"},
-		}}
-	case JaegerID:
-		return map[string]any{"protocols": map[string]any{
-			"grpc":           &ServerConfig{Endpoint: "test", TLSSetting: c.buildTLSSetting()},
-			"thrift_binary":  &ServerConfig{Endpoint: "test:6832"},
-			"thrift_compact": &ServerConfig{Endpoint: "test:6831"},
-			"thrift_http":    &ServerConfig{Endpoint: "test:14268", TLSSetting: c.buildTLSSetting()},
-		}}
-	case ZipkinID:
-		return &ServerConfig{
-			Endpoint:   "test",
-			TLSSetting: c.buildTLSSetting(),
-		}
-	case StatsdID:
-		return map[string]any{
-			"endpoint": "test",
-			"timer_histogram_mapping": []TimerHistogramMapping{{
-				StatsDType: "histogram", ObserverType: "histogram", Histogram: HistogramConfig{MaxSize: 10},
-			}},
-		}
-	}
-
-	return nil
-}
-
-func (c *Config) buildReceivers(protocols []string) (map[component.ID]component.Config, error) {
-	if len(protocols) == 0 {
-		// means all protocols are enabled
-		protocols = []string{string(StatsdProtocol), string(ZipkinProtocol), string(JaegerProtocol), string(OtlpProtocol)}
-	}
-
-	receivers := make(map[component.ID]component.Config)
-
-	for _, p := range protocols {
-		switch Protocol(p) {
-		case StatsdProtocol:
-			receivers[StatsdID] = c.buildReceiverComponent(StatsdID)
-		case ZipkinProtocol:
-			receivers[ZipkinID] = c.buildReceiverComponent(ZipkinID)
-		case JaegerProtocol:
-			receivers[JaegerID] = c.buildReceiverComponent(JaegerID)
-		case OtlpProtocol:
-			receivers[OtlpID] = c.buildReceiverComponent(OtlpID)
-		default:
-			return nil, fmt.Errorf("unknown protocol: %s", p)
-		}
-	}
-
-	return receivers, nil
 }
 
 func WithProtocols(protocols ...string) Option {
