+++ pkg/otelcgen/testdata/receivers_zipkin_only.yaml
@@ -4,7 +4,7 @@ extensions: {}
 processors: {}
 receivers:
     zipkin:
-        endpoint: test
+        endpoint: test:9411
         tls:
             cert_file: /run/opensignals/tls/tls.crt
             key_file: /run/opensignals/tls/tls.key
